+++ recipes/eval/cfg/campaign/dlwp_2024_monthly.yaml
@@ -19,7 +19,15 @@ start_times:
     - "2024-11-01 00:00:00"
     - "2024-12-01 00:00:00"
 
-nsteps: 40
+nsteps: 4
+
+# Cache the full forecast-window valid times into data.zarr so score.py can
+# use it as the verification store.  Must be set here (not just on the
+# predownload.py CLI) because score.py re-reads this block to decide whether
+# data.zarr holds verification data.
+predownload:
+    verification:
+        enabled: true
 
 output:
     variables:
